[refactor] Moved "exceptionAlert" to its own TS file
Since this is a UI only function, when Web Workers are introduced this
call chain will not be allowed on the backend.
