+++ src/CinematicText.js
@@ -1,7 +1,8 @@
 import {Engine}                             from "./engine";
 
 import {removeChildrenFromElement,
-        createElement, exceptionAlert}      from "../utils/HelperFunctions";
+        createElement}                      from "../utils/HelperFunctions";
+import {exceptionAlert}                     from "../utils/helpers/exceptionAlert";
 import {isString}                           from "../utils/helpers/isString";
 
 var cinematicTextFlag = false;
