+++ src/Prestige.js
@@ -31,7 +31,8 @@ import {Terminal, postNetburnerText}            from "./Terminal";
 import Decimal                                  from "decimal.js";
 import {dialogBoxCreate}                        from "../utils/DialogBox";
 import {createPopup, createElement,
-        removeElementById, exceptionAlert}      from "../utils/HelperFunctions";
+        removeElementById}                      from "../utils/HelperFunctions";
+import {exceptionAlert}                         from "../utils/helpers/exceptionAlert";
 import {yesNoBoxCreate, yesNoBoxGetYesButton,
         yesNoBoxGetNoButton, yesNoBoxClose}     from "../utils/YesNoBox";
